+++ wheat-macros/src/main/scala/x7c1/wheat/macros/intent/ServiceCaller.scala
@@ -20,7 +20,7 @@ private object ServiceCallerImpl {
     (context: c.Tree, klass: c.Tree)(f: c.Tree): c.Tree = {
     import c.universe._
 
-    val factory = new ServiceCallerTreeFactory {
+    val factory = new IntentTreeFactory {
       override val context: c.type = c
       override val block = f
     }
@@ -35,48 +35,6 @@ private object ServiceCallerImpl {
 //    println(showCode(tree))
     tree
   }
-}
-
-private trait ServiceCallerTreeFactory {
-  val context: blackbox.Context
-  import context.universe._
-  val block: Tree
-
-  case class IntentExtra(
-    key: String,
-    value: Tree,
-    typeName: String
-  )
-  private lazy val (_, call) = block.children match {
-    case Seq(x, y: Apply) => x -> y
-    case _ => throw new IllegalArgumentException("invalid form of expression")
-  }
-  lazy val intent: TermName = TermName(context freshName "intent")
-
-  lazy val methodName: String = call.symbol.asMethod.fullName
-
-  private def extras: List[IntentExtra] = {
-    val pairs = call.symbol.asMethod.paramLists match {
-      case xs if xs.length > 1 =>
-        throw new IllegalArgumentException(s"too many paramLists : $xs")
-      case Seq(params) =>
-        params map { param =>
-          param.name.encodedName.toString ->
-            param.typeSignature.typeSymbol.fullName
-        }
-    }
-    pairs zip call.children.tail map {
-      case ((paramName, paramType), arg) =>
-        IntentExtra(
-          key = s"$methodName.$paramName",
-          value = arg,
-          typeName = paramType )
-    }
-  }
-  def putExtras(intent: TermName): List[Tree] = extras map { extra =>
-    q"$intent.putExtra(${extra.key}, ${extra.value})"
-  }
-
 }
 
 trait BundleConvertible[A] {
